Fix dep loop by moving route based helps in auth out to utils file
